refactor: extract OpenAI-compatible chat-completions transport class (#25846)
* refactor: extract OpenAI-compatible chat-completions transport class

* fix: update web-search guard allowlist for renamed chat-completions-provider file
