+++ assistant/src/__tests__/openai-provider.test.ts
@@ -89,6 +89,7 @@ mock.module("openai", () => ({
 // Import after mocking
 import { FireworksProvider } from "../providers/fireworks/client.js";
 import { OllamaProvider } from "../providers/ollama/client.js";
+import { OpenAIChatCompletionsProvider } from "../providers/openai/chat-completions-provider.js";
 import { OpenAIProvider } from "../providers/openai/client.js";
 import { OpenRouterProvider } from "../providers/openrouter/client.js";
 
@@ -181,6 +182,32 @@ function reasoningUsageChunk(
     model: "gpt-5.2",
   };
 }
+
+// ---------------------------------------------------------------------------
+// Class extraction sanity checks
+// ---------------------------------------------------------------------------
+
+describe("OpenAIChatCompletionsProvider extraction", () => {
+  test("OpenAIProvider is an alias for OpenAIChatCompletionsProvider", () => {
+    expect(OpenAIProvider).toBe(OpenAIChatCompletionsProvider);
+  });
+
+  test("compatibility providers extend OpenAIChatCompletionsProvider", () => {
+    lastConstructorOptions = null;
+
+    const or = new OpenRouterProvider("or-key", "openai/gpt-4o");
+    expect(or).toBeInstanceOf(OpenAIChatCompletionsProvider);
+
+    const fw = new FireworksProvider(
+      "fw-key",
+      "accounts/fireworks/models/llama-v3p1-70b-instruct",
+    );
+    expect(fw).toBeInstanceOf(OpenAIChatCompletionsProvider);
+
+    const ol = new OllamaProvider("llama3.2");
+    expect(ol).toBeInstanceOf(OpenAIChatCompletionsProvider);
+  });
+});
 
 // ---------------------------------------------------------------------------
 // Tests
